// core/types/app-state.ts
import {GradeOptState} from '../logic/grade-opt/reducer';
import {LanguageState} from '../models/app-language';
import {ModuleCollection} from '../models/credit';
import {CoronaState} from '../reducers/corona';
import {CountsTowardsCreditsState} from '../reducers/countsTowardsCredits';
import {CreditOverrideState} from '../reducers/creditOverrides';
import {CustomCreditsState} from '../reducers/customCredits';
import {ExamReturnState} from '../reducers/exam-returns';
import {HiddenContentState} from '../reducers/hiddenContent';
import {RoomState} from '../reducers/room';
import {ScheduleReducerState} from '../reducers/ScheduleReducerState';
import {SelectedMessagesState} from '../reducers/selected-messages';
import {SeriesConfigState} from '../reducers/seriesConfig';
import {SyncState} from '../reducers/sync';
import {NotificationsState} from '../types/notifications-state';
import {ApiReducerState} from './api-reducer-state';
import {AppearanceState} from './appearance-state';
import {ChatServerState} from './chat';
import {ConnectivityReducerState} from './connectivity-state';
import {CountsTowardsAverageState} from './counts-towards-average-state';
import {AppFoodState} from './food';
import {GradeStatisticsReducerType} from './grade-statistics';
import {InstitutionState} from './institution-state';
import {LoginState} from './login-state';
import {ModalsState} from './modals';
import {ModuleRatingsState} from './module-ratings-state';
import {PeopleState} from './people-state';
import {PromotionState} from './promotion-state';
import {MyRatingsState, RatingsState} from './ratings';
import {ReadyState} from './ready-state';
import {SearchState} from './search-state';
import {TimeTableState} from './timetable-state';
import {UserState} from './user-state';
import {UzhLoginResponse} from './uzh-login';

export type AppState = {
	api: ApiReducerState;
	countsTowardsAverage: CountsTowardsAverageState;
	countsTowardsCredits: CountsTowardsCreditsState;
	schedule: ScheduleReducerState;
	connectivity: ConnectivityReducerState;
	moduleCollection: ModuleCollection;
	people: PeopleState;
	multiLogin: {
		UZH: LoginState;
		ETH: LoginState;
	};
	multiSummary: {
		UZH: UzhLoginResponse | null;
		ETH: UzhLoginResponse | null;
	};
	multiGradeStatistics: {
		UZH: GradeStatisticsReducerType;
		ETH: GradeStatisticsReducerType;
	};
	timetable: TimeTableState;
	institution: InstitutionState;
	hiddenContent: HiddenContentState;
	ready: ReadyState;
	moduleRatings: ModuleRatingsState;
	creditOverrides: CreditOverrideState;
	multiMyRatings: {
		UZH: MyRatingsState;
		ETH: MyRatingsState;
	};
	search: SearchState;
	seriesConfig: SeriesConfigState;
	promotions: PromotionState;
	food: AppFoodState;
	ratings: RatingsState;
	room: RoomState;
	modals: ModalsState;
	chatServer: ChatServerState;
	users: UserState;
	notifications: NotificationsState;
	language: LanguageState;
	appearance: AppearanceState;
	customCredits: CustomCreditsState;
	examReturns: ExamReturnState;
	corona: CoronaState;
	gradeOpt: GradeOptState;
	sync: SyncState;
	selectedMessages: SelectedMessagesState;
};